+++ DK/DKFramework/Private/Cocoa/Window.mm
@@ -372,7 +372,7 @@ void Window::Resize(DKSize s, const DKPoint* pt)
 	});
 }
 
-double Window::ContentScaleFactor() const
+float Window::ContentScaleFactor() const
 {
 	return DispatchSyncOnMain([&]()->CGFloat {
 		if (view.window)
